requirements.txt
tensorflow
numpy
scikit-learn
pandas
matplotlib
tqdm
biopython